automation-system/requirements.txt
requests==2.31.0
flask==3.0.3
schedule==1.2.2
python-dotenv==1.0.1
pyyaml==6.0.1
linebot==3.10.1
anthropic>=0.28.0
# Google Drive 連携
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.126.0
# 本番サーバー
gunicorn==22.0.0
